qr-code-master/requirements.txt
altgraph==0.17
colorama==0.4.4
future==0.18.2
image==1.2.1
imageio==2.9.0
importlib-metadata==4.6.1
MyQR==2.3.1
numpy==1.21.1
opencv-python==4.5.3.56
pefile==2021.5.24
Pillow==8.3.1
pycparser==2.20
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyserial==3.5
pythonnet==2.5.2
pyzbar==0.1.8
qrcode==7.2
typing-extensions==3.10.0.0
WMI==1.5.1
zipp==3.5.0
